only do completion_queue in addition to legacy
